Combine clear-schedule and schedule into scheduler command
Instead of using `--append` to signify when to append or clear, add
`--clear`. Also, schedule must be provided via `--schedule`. So if
`--schedule` is provided, the events will be added to exiting ones.
If `--clear` is also added, events will be cleared before new ones are
added. And if `--clear` is provided without `--schedule`, events will
simply be cleared. If nothing is provided, nothing will be done.